refactor: separate platform specific logic into separate files
