+++ src/common/os_utils/src/os_utils_unix.cpp
@@ -9,25 +9,15 @@
 #include <sstream>
 #include <unistd.h>
 
-#if defined(__APPLE__)
-#include <libproc.h>
-#elif defined(__linux__)
 #include <sys/types.h>
-#endif
 
 std::vector<std::string> os_utils::OsUtils::GetRunningProcesses()
 {
     return os_utils::GetRunningProcesses();
 }
 
-bool os_utils::PidExists(pid_t pid)
-{
-    return !(getsid(pid) == -1 && errno == ESRCH) && !(getpgid(pid) == -1 && errno == ESRCH);
-}
-
 std::string os_utils::GetProcessName(pid_t pid)
 {
-#ifdef __linux__
     const auto commPath = std::filesystem::path("/proc") / std::to_string(pid) / "comm";
 
     if (std::ifstream comm(commPath); comm)
@@ -37,20 +27,7 @@ std::string os_utils::GetProcessName(pid_t pid)
         std::getline(comm, name);
         return name;
     }
-#elif defined(__APPLE__)
-    if (char pathbuf[PROC_PIDPATHINFO_MAXSIZE]; proc_pidpath(pid, pathbuf, sizeof(pathbuf)) > 0)
-    {
-        const std::string fullPath(pathbuf);
 
-        if (const auto pos = fullPath.find_last_of('/'); pos != std::string::npos)
-        {
-            return fullPath.substr(pos + 1);
-        }
-        return fullPath;
-    }
-#else
-#error "Unsupported platform"
-#endif
     return {};
 }
 
@@ -58,7 +35,6 @@ std::vector<std::string> os_utils::GetRunningProcesses()
 {
     std::vector<std::string> processList;
 
-#ifdef __linux__
     for (const auto& entry : std::filesystem::directory_iterator("/proc"))
     {
         const auto& filename = entry.path().filename().string();
@@ -73,20 +49,6 @@ std::vector<std::string> os_utils::GetRunningProcesses()
             }
         }
     }
-#else
-    const pid_t maxPid = 99999;
-
-    for (pid_t pid = 1; pid <= maxPid; ++pid)
-    {
-        if (PidExists(pid))
-        {
-            if (auto name = GetProcessName(pid); !name.empty())
-            {
-                processList.emplace_back(std::move(name));
-            }
-        }
-    }
-#endif
 
     return processList;
 }
